EDM-3145: adding flightctl restore to downloadable CLIs
