+++ internal/service/device.go
@@ -23,87 +23,6 @@ import (
 	"github.com/samber/lo"
 )
 
-// PrepareDevicesAfterRestore performs post-restoration preparation tasks for devices
-func (h *ServiceHandler) PrepareDevicesAfterRestore(ctx context.Context) error {
-	h.log.Info("Starting post-restoration device preparation")
-
-	// 1. Drop the KV store to clear all cached data
-	h.log.Info("Clearing KV store after restoration")
-	if h.kvStore != nil {
-		if err := h.kvStore.DeleteAllKeys(ctx); err != nil {
-			h.log.WithError(err).Error("Failed to clear KV store")
-			return fmt.Errorf("failed to clear KV store: %w", err)
-		}
-		h.log.Info("KV store cleared successfully")
-	} else {
-		h.log.Warn("KV store not available, skipping clear")
-	}
-
-	// 2. Set waitForDeviceToReconnectAfterRestore annotation on all devices and unset lastSeen
-	h.log.Info("Updating device annotations and clearing lastSeen timestamps")
-
-	devicesUpdated, err := h.store.Device().PrepareDevicesAfterRestore(ctx)
-	if err != nil {
-		h.log.WithError(err).Error("Failed to prepare devices after restore")
-		return fmt.Errorf("failed to prepare devices after restore: %w", err)
-	}
-
-	// 3. Set awaitingReconnection annotation on all non-approved enrollment requests
-	h.log.Info("Updating enrollment request annotations for non-approved requests")
-
-	enrollmentRequestsUpdated, err := h.store.EnrollmentRequest().PrepareEnrollmentRequestsAfterRestore(ctx)
-	if err != nil {
-		h.log.WithError(err).Error("Failed to prepare enrollment requests after restore")
-		return fmt.Errorf("failed to prepare enrollment requests after restore: %w", err)
-	}
-
-	// 4. Add awaiting reconnection keys for all devices across all organizations
-	organizations, err := h.store.Organization().List(ctx, store.ListParams{})
-	if err != nil {
-		h.log.WithError(err).Error("Failed to get organizations for awaiting reconnection keys")
-		return fmt.Errorf("failed to get organizations for awaiting reconnection keys: %w", err)
-	}
-	h.log.Infof("Adding awaiting reconnection keys for %d organizations", len(organizations))
-
-	awaitingReconnectionKeysAdded := 0
-	for _, org := range organizations {
-		deviceNames, err := h.store.Device().GetAllDeviceNames(ctx, org.ID)
-		if err != nil {
-			h.log.WithError(err).Errorf("Failed to get device names for organization %s", org.ID)
-			// Continue with other organizations even if one fails
-			continue
-		}
-
-		for _, deviceName := range deviceNames {
-			key := kvstore.AwaitingReconnectionKey{
-				OrgID:      org.ID,
-				DeviceName: deviceName,
-			}
-			keyStr := key.ComposeKey()
-			_, err := h.kvStore.SetNX(ctx, keyStr, []byte("true"))
-			if err != nil {
-				h.log.WithError(err).Errorf("Failed to add awaiting reconnection key for device %s in org %s", deviceName, org.ID)
-				// Continue with other devices even if one fails
-				continue
-			}
-			awaitingReconnectionKeysAdded++
-		}
-	}
-
-	h.log.Infof("Post-restoration device preparation completed successfully. Updated %d devices, %d enrollment requests. Added %d awaiting reconnection keys across %d organizations.", devicesUpdated, enrollmentRequestsUpdated, awaitingReconnectionKeysAdded, len(organizations))
-
-	// Emit system restored event
-	if h.eventHandler != nil {
-		event := common.GetSystemRestoredEvent(ctx, devicesUpdated)
-		if event != nil {
-			// No org available in this context; emit under NullOrgId explicitly
-			h.eventHandler.CreateEvent(ctx, store.NullOrgId, event)
-			h.log.Info("System restored event created successfully")
-		}
-	}
-	return nil
-}
-
 func (h *ServiceHandler) CreateDevice(ctx context.Context, orgId uuid.UUID, device domain.Device) (*domain.Device, domain.Status) {
 	if device.Spec != nil && device.Spec.Decommissioning != nil {
 		h.log.WithError(flterrors.ErrDecommission).Error("attempt to create decommissioned device")
