+++ test/integration/service/enrollmentrequest_awaiting_reconnect_test.go
@@ -205,91 +205,5 @@ var _ = Describe("EnrollmentRequest AwaitingReconnect Integration Tests", func()
 		})
 	})
 
-	Context("PrepareDevicesAfterRestore integration", func() {
-		It("should annotate enrollment requests and create devices with awaitingReconnect status", func() {
-			// Create multiple enrollment requests with different states
-			er1 := CreateTestER()
-			er1Name := lo.FromPtr(er1.Metadata.Name)
-
-			er2 := CreateTestER()
-			er2Name := lo.FromPtr(er2.Metadata.Name)
-
-			er3 := CreateTestER()
-			er3Name := lo.FromPtr(er3.Metadata.Name)
-
-			By("creating enrollment requests")
-			// Use internal request context to preserve annotations
-			internalCtx := context.WithValue(suite.Ctx, consts.InternalRequestCtxKey, true)
-			_, status := suite.Handler.CreateEnrollmentRequest(internalCtx, suite.OrgID, er1)
-			Expect(status.Code).To(BeEquivalentTo(http.StatusCreated))
-
-			_, status = suite.Handler.CreateEnrollmentRequest(internalCtx, suite.OrgID, er2)
-			Expect(status.Code).To(BeEquivalentTo(http.StatusCreated))
-
-			_, status = suite.Handler.CreateEnrollmentRequest(internalCtx, suite.OrgID, er3)
-			Expect(status.Code).To(BeEquivalentTo(http.StatusCreated))
-
-			By("approving one enrollment request before restore")
-			defaultOrg := &model.Organization{
-				ID:          org.DefaultID,
-				ExternalID:  org.DefaultID.String(),
-				DisplayName: org.DefaultID.String(),
-			}
-			mappedIdentity := identity.NewMappedIdentity("testuser", "", []*model.Organization{defaultOrg}, map[string][]string{}, false, nil)
-			ctxApproval := context.WithValue(suite.Ctx, consts.MappedIdentityCtxKey, mappedIdentity)
-
-			approval := api.EnrollmentRequestApproval{
-				Approved: true,
-				Labels:   &map[string]string{"approved": "true"},
-			}
-
-			_, st := suite.Handler.ApproveEnrollmentRequest(ctxApproval, suite.OrgID, er1Name, approval)
-			Expect(st.Code).To(BeEquivalentTo(http.StatusOK))
-
-			By("simulating restore process - annotating non-approved enrollment requests")
-			enrollmentRequestsUpdated, err := suite.Store.EnrollmentRequest().PrepareEnrollmentRequestsAfterRestore(suite.Ctx)
-			Expect(err).ToNot(HaveOccurred())
-			Expect(enrollmentRequestsUpdated).To(Equal(int64(2)), "Should update 2 non-approved enrollment requests")
-
-			By("verifying enrollment requests were annotated")
-			er2Updated, status := suite.Handler.GetEnrollmentRequest(suite.Ctx, suite.OrgID, er2Name)
-			Expect(status.Code).To(BeEquivalentTo(http.StatusOK))
-			Expect(er2Updated.Metadata.Annotations).ToNot(BeNil())
-			Expect(*er2Updated.Metadata.Annotations).To(HaveKeyWithValue(api.DeviceAnnotationAwaitingReconnect, "true"))
-
-			er3Updated, status := suite.Handler.GetEnrollmentRequest(suite.Ctx, suite.OrgID, er3Name)
-			Expect(status.Code).To(BeEquivalentTo(http.StatusOK))
-			Expect(er3Updated.Metadata.Annotations).ToNot(BeNil())
-			Expect(*er3Updated.Metadata.Annotations).To(HaveKeyWithValue(api.DeviceAnnotationAwaitingReconnect, "true"))
-
-			// er1 should not have the annotation since it was already approved
-			er1Updated, status := suite.Handler.GetEnrollmentRequest(suite.Ctx, suite.OrgID, er1Name)
-			Expect(status.Code).To(BeEquivalentTo(http.StatusOK))
-			Expect(er1Updated.Metadata.Annotations).ToNot(BeNil())
-			Expect(*er1Updated.Metadata.Annotations).To(BeEmpty())
-
-			By("approving the annotated enrollment requests and verifying device creation")
-			_, st = suite.Handler.ApproveEnrollmentRequest(ctxApproval, suite.OrgID, er2Name, approval)
-			Expect(st.Code).To(BeEquivalentTo(http.StatusOK))
-
-			_, st = suite.Handler.ApproveEnrollmentRequest(ctxApproval, suite.OrgID, er3Name, approval)
-			Expect(st.Code).To(BeEquivalentTo(http.StatusOK))
-
-			By("verifying devices were created with awaitingReconnect status")
-			device2, status := suite.Handler.GetDevice(suite.Ctx, suite.OrgID, er2Name)
-			Expect(status.Code).To(BeEquivalentTo(http.StatusOK))
-			Expect(device2).ToNot(BeNil())
-			Expect(device2.Metadata.Annotations).ToNot(BeNil())
-			Expect(*device2.Metadata.Annotations).To(HaveKeyWithValue(api.DeviceAnnotationAwaitingReconnect, "true"))
-			Expect(device2.Status.Summary.Status).To(Equal(api.DeviceSummaryStatusAwaitingReconnect))
-
-			device3, status := suite.Handler.GetDevice(suite.Ctx, suite.OrgID, er3Name)
-			Expect(status.Code).To(BeEquivalentTo(http.StatusOK))
-			Expect(device3).ToNot(BeNil())
-			Expect(device3.Metadata.Annotations).ToNot(BeNil())
-			Expect(*device3.Metadata.Annotations).To(HaveKeyWithValue(api.DeviceAnnotationAwaitingReconnect, "true"))
-			Expect(device3.Status.Summary.Status).To(Equal(api.DeviceSummaryStatusAwaitingReconnect))
-
-		})
-	})
+	// PrepareDevicesAfterRestore integration tests have been moved to test/integration/restore/enrollmentrequest_test.go
 })
